Add annotating, inferring type checker.
The type checker infers types for all expressions and annotates the parse tree
with type variables. The annotated tree and active substitution is then
returned so that the code generator will be able to make use of the types.

It does not yet do kind checking (which gives even worse error messages, but
at least doesn't allow any illegal programs through, where illegal is defined
as "does the wrong thing at runtime") or explicit type declarations.