thunk function arguments, and eval all identifiers. Cut all other lazy stuff.
Laziness is ONLY introduced when binding a value to a symbol. Symbols are
eval'd whenever they're used. This commit removes the old random thunk/eval
and introduces evaluation for all symbols, and thunks all function arguments.

The observant reader will notice that this breaks global and local function
definitions, since this commit does NOT thunk those. That's not a big deal
though, as we've yet to produce any working code, local function definitions
aren't even implemented in the codegen, and those issues will be resolved
promptly.