Create proper LS list-of-char strings, and print errors to stdout.
The codegen now asks the runtime to turn string literals into list-of-char
strings, and the error function turns them back into JS strings again before
printing.
Errors are now also printed to stdout before an exception is thrown. This is
to give test cases a facility for reporting string results until we get IO
done.
The case.ls test case was slightly modified to take advantage of this.