Pass max. # of function args through to the optimizer.
If we're to inline function calls properly (inline, in this context, means
turning function(a) {return function(b) {...}} and f(a)(b) into
function(a, b) {...} and f(a, b) respectively) the optimizer needs to know
if a function is being completely or partially applied. This commit passes
this information through to the optimizer, but doesn't do anything with it
yet.