Add function inlining to the list of optimizations, but don't enable it.
This commit introduces an optimization that turns \a -> \b -> ... expressions
into \a b -> ... expressions. Since these are very different in Javascript
(the first is called like f(a)(b) and the second like f(a,b)) this
optimization will break just about everything if enabled without the
supporting FoldCalls optimization, that turns f(a)(b) calls into f(a,b).

Incidentally, FoldCalls has yet to be written.