Inline functions returning functions into multi-arg function calls.
We now generate code like f(a,b,c) for function calls instead of f(a)(b)(c).
In the case of a partially applied function, a JS closure is generated.
This commit also introduces a bug, which causes incorrect code to be generated
when a function argument is partially applied in a generic function. Assume:
f :: (a -> b) -> a -> b;
f g x = g x;

g :: Int -> Int -> Int;
g a b = a + b;

h = (f g 1) 1;

Here, even though g really has the type Int -> Int -> Int, in its role as an
argument to function f it has the type a -> b. This means that f is unable
to generate the call g x as function(y) {return g(x, y);} as it should and
instead spits out g(x) - as far as f is concerned, g takes one argument and
gives one output.

The solution to this problem is to examine the type of a function's formal
parameters when passing a function as an argument and, if the arity of the
formal and actual parameter doesn't match up, pass a closure instead:
f(function(x) {return function(y) {return g(x, y);}}, 1)
Since the return type of g is completely opaque to f it won't call g with
more than one argument (f doesn't know you can even do that!) and the
resulting partial application will be exactly what everyone else expects.

This requires a few changes to codegen (namely, pass along type info for
function args) and so will be introduced in another series of commits.