Thunkify global nullary functions.
While there is no benefit in having to evaluate a thunk to get at a function
of arity 1 or above, keeping nullary functions (that is, constants) as thunks
can be rather nice. Especially since it doesn't add much overhead; the
indirection is already there, the only difference is one tiny if, with
potentially large benefits if calculating the value of the constant is
expensive.