Desugar do { a ; b ; ... }; into a >>= b >>= ...
Later on we'll be turning this sort of bind chain into purely imperative code,
but the cleanest way to type check this sort of thing is to turn it into
something the type checker already understands. As a bonus, it allows us to
get (slow) IO up and running pretty fast since all we need for that now are
proper implementations of bind and return.