Inline calls to _jsfun where the first argument is a string literal.
Since 99% of all calls to _jsfun, and 100% of all _sane_ calls to it, will
use a string literal as its first argument, there's really no point in
converting that string to a list-of-char string and back again.

It's still possible to use an expression other than a literal with _jsfun,
but it's going to be MUCH slower than using a literal. (It's also horrible
abuse of _jsfun so please don't do it.)