Fix compatibility with Opera, Chrome.
The arity attribute is apparently deprecated in modern JS; instead we have to
use the length attribute. Because it makes perfect sense that if a function
takes 3 arguments, then it's obviously 3 long.

Also, since return (and others) is a reserved keyword in JS, and
implementations other than Mozilla's respect that, we can't do $.return = ...
Instead, we do $['return'] = ...; it's up to the user not to use any reserved
keywords in their export names. If they really want to, then they can do
module['reserved__word'](...).