Compare to undefined instead using typeof to check if a thunk is evaluated.
Previously we called a function _u(o) that did typeof o.x == 'undefined' to
check if the thunk o is previously evaluated or not. Not only does that
require a staggering amount of function calls, but also includes expensive
string comparisons unless the JS interpreter does a good job optimizing that
particular idiom. Comparing an attribute to undefined should be much faster,
and brings the total number of function calls for the Tetris example down
to "only" 1/3 million per second (from the old ~1.1M.)