Make sure that conditions in ifs and cases are never evaluated too early.
Previously, any expression the code generation of which also resulted in at
least one statement being generated would have that statement executed
before the thunk was evaluated. This is OK for local function bindings where
the entire function is thunked anyway, but the other statement users, ifs and
cases, would have their condition prematurely evaluated in some cases.

This commit fixes the issue by creating a closure to run execute the
statements inside. This means that instead of prematurely executing code and
then assigning the result to a variable, we can include the entire bunch of
statements in an expression instead.

This is probably a bit slower, but it should be really easy to inline in most
cases, if we ever get around to doing that.