Fix imported higher order functions.
Normally, calling a function of arity n takes exactly n function calls. For
nullary functions, however, we still use one call. This requires some special
handling in the optimizer, which needs to know the arity of global functions.

This information wasn't provided for imported functions before, and the
optimizer didn't even know about them, which led to crashes when using an
imported function as a higher order function.