Codegen isn't buggy for local functions; a Firefox bug is the culprit.
The bug in question manifests itself in certain circumstances when a variable
that should be bound in a closure suddenly gets unbound. An example of this
happening:
var _11 = function(){
  return function(_12){
  alert("_12 is bound: " + _12);
  return (function(){
    alert("_12 is NOT bound! " + _12);
    if((_14=_12())[0] == 1) {
      var _15 = _14[1];
      return _15();
    }
    return error(function _(){if(_.x===$u) _.x=_s("non-exhaustive pattern in function!");return _.x;});
  })();
};

Unfortunately, just calling _11(function() {return [1, undefined];}) does NOT
reproduce the problem so the bug seems to be affected by the circumstances
surrounding the actual call.