Revamp type checking dependent modules to disable using an import's imports.
Previously the type checking of one module would take as its input the set of
assumptions and types to start with, and return those assumptions and types
plus whatever the type checker felt like adding. This worked sort of OK mostly
but was really a total mess. As of this commit, exported types and functions
are properly sorted by exporting module, and no module gets to see exports
from modules it didn't explicitly ask for.