Eliminate all tailcalls to non-nullary functions by trampolining.
As of this commit, every function is entered through a trampoline. A tail call
is made by returning {f:<the function expression>, a:<the argument>} to the
trampoline rather than a "normal" return value. This makes the trampoline
fairly non-intrusive, so code compiled with TCE disabled works just fine
without having to do any voodoo to the runtime libraries.